# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shift — AI-powered visitor personalization engine
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Visitor {
  id: number;
  sessionId: string;
  /** @nullable */
  referrer?: string | null;
  /** @nullable */
  utmSource?: string | null;
  /** @nullable */
  utmMedium?: string | null;
  /** @nullable */
  utmCampaign?: string | null;
  /** @nullable */
  userAgent?: string | null;
  /** @nullable */
  deviceType?: string | null;
  /** @nullable */
  intentSignals?: string | null;
  persona: string;
  /** @nullable */
  personaConfidence?: number | null;
  /** @nullable */
  headline?: string | null;
  /** @nullable */
  subheadline?: string | null;
  /** @nullable */
  ctaText?: string | null;
  converted: boolean;
  /** @nullable */
  conversionEvent?: string | null;
  /** @nullable */
  timeOnSite?: number | null;
  createdAt: string;
  updatedAt?: string;
}

export interface VisitorInput {
  sessionId: string;
  referrer?: string;
  utmSource?: string;
  utmMedium?: string;
  utmCampaign?: string;
  userAgent?: string;
  deviceType?: string;
}

export interface VisitorUpdate {
  converted?: boolean;
  conversionEvent?: string;
  timeOnSite?: number;
  persona?: string;
  headline?: string;
  subheadline?: string;
  ctaText?: string;
}

export interface IntentInput {
  sessionId: string;
  referrer?: string;
  utmSource?: string;
  utmMedium?: string;
  utmCampaign?: string;
  userAgent?: string;
  deviceType?: string;
  pageTitle?: string;
}

export interface IntentResult {
  persona: string;
  confidence: number;
  headline: string;
  subheadline: string;
  ctaText: string;
  funnelTheme: string;
  reasoning?: string;
  /** @nullable */
  visitorId?: number | null;
}

export interface DashboardStats {
  totalVisitors: number;
  totalConverted: number;
  conversionRate: number;
  uniquePersonas: number;
  /** @nullable */
  avgTimeOnSite: number | null;
  todayVisitors?: number;
  todayConverted?: number;
}

export interface FunnelStat {
  persona: string;
  count: number;
  converted: number;
  conversionRate: number;
  /** @nullable */
  avgConfidence?: number | null;
}

export type ListVisitorsParams = {
limit?: number;
offset?: number;
persona?: string;
};

export type GetRecentActivityParams = {
limit?: number;
};