# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shift — AI-powered visitor personalization engine
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalVisitors: number;
  totalConverted: number;
  conversionRate: number;
  uniquePersonas: number;
  /** @nullable */
  avgTimeOnSite: number | null;
  todayVisitors?: number;
  todayConverted?: number;
}