# lib/api-zod/src/generated/types/funnelStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shift — AI-powered visitor personalization engine
 * OpenAPI spec version: 0.1.0
 */

export interface FunnelStat {
  persona: string;
  count: number;
  converted: number;
  conversionRate: number;
  /** @nullable */
  avgConfidence?: number | null;
}