# lib/api-zod/src/generated/types/intentResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shift — AI-powered visitor personalization engine
 * OpenAPI spec version: 0.1.0
 */

export interface IntentResult {
  persona: string;
  confidence: number;
  headline: string;
  subheadline: string;
  ctaText: string;
  funnelTheme: string;
  reasoning?: string;
  /** @nullable */
  visitorId?: number | null;
}