# lib/api-zod/src/generated/types/visitor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shift — AI-powered visitor personalization engine
 * OpenAPI spec version: 0.1.0
 */

export interface Visitor {
  id: number;
  sessionId: string;
  /** @nullable */
  referrer?: string | null;
  /** @nullable */
  utmSource?: string | null;
  /** @nullable */
  utmMedium?: string | null;
  /** @nullable */
  utmCampaign?: string | null;
  /** @nullable */
  userAgent?: string | null;
  /** @nullable */
  deviceType?: string | null;
  /** @nullable */
  intentSignals?: string | null;
  persona: string;
  /** @nullable */
  personaConfidence?: number | null;
  /** @nullable */
  headline?: string | null;
  /** @nullable */
  subheadline?: string | null;
  /** @nullable */
  ctaText?: string | null;
  converted: boolean;
  /** @nullable */
  conversionEvent?: string | null;
  /** @nullable */
  timeOnSite?: number | null;
  createdAt: string;
  updatedAt?: string;
}