# lib/api-zod/src/generated/types/visitorInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shift — AI-powered visitor personalization engine
 * OpenAPI spec version: 0.1.0
 */

export interface VisitorInput {
  sessionId: string;
  referrer?: string;
  utmSource?: string;
  utmMedium?: string;
  utmCampaign?: string;
  userAgent?: string;
  deviceType?: string;
}